Move the test for log_size to its own file
